+++ eyed3/core.py
@@ -91,7 +91,7 @@ def load(path, tag_version=None):
         return None
 
 
-class AudioInfo(object):
+class AudioInfo:
     """A base container for common audio details."""
 
     # The number of seconds of audio data (i.e., the playtime)
@@ -100,7 +100,7 @@ class AudioInfo(object):
     size_bytes = 0
 
 
-class Tag(object):
+class Tag:
     """An abstract interface for audio tag (meta) data (e.g. artist, title,
     etc.)
     """
@@ -108,34 +108,34 @@ class Tag(object):
     read_only = False
 
     def _setArtist(self, val):
-        raise NotImplementedError
+        raise NotImplementedError()  # pragma: nocover
 
     def _getArtist(self):
-        raise NotImplementedError
+        raise NotImplementedError()  # pragma: nocover
 
     def _getAlbumArtist(self):
-        raise NotImplementedError
+        raise NotImplementedError()  # pragma: nocover
 
     def _setAlbumArtist(self, val):
-        raise NotImplementedError
+        raise NotImplementedError()  # pragma: nocover
 
     def _setAlbum(self, val):
-        raise NotImplementedError
+        raise NotImplementedError()  # pragma: nocover
 
     def _getAlbum(self):
-        raise NotImplementedError
+        raise NotImplementedError()  # pragma: nocover
 
     def _setTitle(self, val):
-        raise NotImplementedError
+        raise NotImplementedError()  # pragma: nocover
 
     def _getTitle(self):
-        raise NotImplementedError
+        raise NotImplementedError()  # pragma: nocover
 
     def _setTrackNum(self, val):
-        raise NotImplementedError
+        raise NotImplementedError()  # pragma: nocover
 
     def _getTrackNum(self):
-        raise NotImplementedError
+        raise NotImplementedError()  # pragma: nocover
 
     @property
     def artist(self):
@@ -181,8 +181,7 @@ class Tag(object):
     def track_num(self, v):
         self._setTrackNum(v)
 
-    def __init__(self, title=None, artist=None, album=None, album_artist=None,
-                 track_num=None):
+    def __init__(self, title=None, artist=None, album=None, album_artist=None, track_num=None):
         self.title = title
         self.artist = artist
         self.album = album
@@ -190,7 +189,7 @@ class Tag(object):
         self.track_num = track_num
 
 
-class AudioFile(object):
+class AudioFile:
     """Abstract base class for audio file types (AudioInfo + Tag)"""
 
     def _read(self):
@@ -234,10 +233,11 @@ class AudioFile(object):
         return self._path
 
     @path.setter
-    def path(self, t):
+    def path(self, path):
         """Set the path"""
-        from os.path import abspath, realpath, normpath
-        self._path = normpath(realpath(abspath(t)))
+        if isinstance(path, pathlib.Path):
+            path = str(path)
+        self._path = path
 
     @property
     def info(self):
@@ -267,7 +267,7 @@ class AudioFile(object):
 
 
 @functools.total_ordering
-class Date(object):
+class Date:
     """
     A class for representing a date and time (optional). This class differs
     from ``datetime.datetime`` in that the default values for month, day,