+++ test/id3/test_tag.py
@@ -1,7 +1,9 @@
 import os
 import pytest
 import unittest
-from pathlib import Path
+
+import deprecation
+
 import eyed3
 from eyed3.core import Date
 from eyed3.id3 import frames
@@ -33,8 +35,6 @@ def testFileInfoConstructor():
         assert type(fi.name) is str
         assert name == str(name)
         assert fi.tag_size == 0
-
-    # FIXME Passing invalid unicode
 
 
 def testTagMainProps():
@@ -123,13 +123,12 @@ def testTagDates():
         tag.tagging_date = str(date)
         assert (tag.tagging_date == date)
 
-
-        try:
-            tag._setDate(2.4)
-        except TypeError:
-            pass # expected
-        else:
-            assert not("Invalid date type, expected TypeError")
+    try:
+        tag._setDate(b"TDRL", 2.4)
+    except TypeError:
+        pass # expected
+    else:
+        assert not("Invalid date type, expected TypeError")
 
 
 def testTagComments():
@@ -766,9 +765,9 @@ def testTagGenre():
     try:
         tag.genre = b"Not Unicode"
     except TypeError:
-        pass # expected
+        pass  # expected
     else:
-        assert not("Non unicode genre, expected TypeError")
+        assert not "Non unicode genre, expected TypeError"
 
     gobj = Genre("Hardcore")
 
@@ -1194,63 +1193,78 @@ def testReadOnly():
         t._saveV2Tag(None, None, None)
 
 
-@pytest.mark.skipif(not Path(DATA_D).exists(), reason="test requires data files")
-def testIssue76(audiofile):
-    """
-    https://github.com/nicfit/eyeD3/issues/76
-    """
-    tag = audiofile.initTag(ID3_V2_4)
-    tag.setTextFrame("TPE1", "Confederacy of Ruined Lives")
-    tag.setTextFrame("TPE2", "Take as needed for pain")
-    tag.setTextFrame("TSOP", "In the name of suffering")
-    tag.setTextFrame("TSO2", "Dope sick")
-    tag.save()
-
-    audiofile = eyed3.load(audiofile.path)
-    tag = audiofile.tag
-    assert (set(tag.frame_set.keys()) ==
-            set([b"TPE1", b"TPE2", b"TSOP", b"TSO2"]))
-    assert tag.getTextFrame("TSO2") == "Dope sick"
-    assert tag.getTextFrame("TSOP") == "In the name of suffering"
-    assert tag.getTextFrame("TPE2") == "Take as needed for pain"
-    assert tag.getTextFrame("TPE1") == "Confederacy of Ruined Lives"
-
-    audiofile.tag.lyrics.set("some lyrics")
-    audiofile = eyed3.load(audiofile.path)
-    tag = audiofile.tag
-    assert (set(tag.frame_set.keys()) ==
-            set([b"TPE1", b"TPE2", b"TSOP", b"TSO2"]))
-    assert tag.getTextFrame("TSO2") == "Dope sick"
-    assert tag.getTextFrame("TSOP") == "In the name of suffering"
-    assert tag.getTextFrame("TPE2") == "Take as needed for pain"
-    assert tag.getTextFrame("TPE1") == "Confederacy of Ruined Lives"
-
-    # Convert to v2.3 and verify conversions
-    tag.save(version=ID3_V2_3)
-    audiofile = eyed3.load(audiofile.path)
-    tag = audiofile.tag
-    assert (set(tag.frame_set.keys()) ==
-            set([b"TPE1", b"TPE2", b"XSOP", b"TSO2"]))
-    assert tag.getTextFrame("TSO2") == "Dope sick"
-    assert tag.getTextFrame("TPE2") == "Take as needed for pain"
-    assert tag.getTextFrame("TPE1") == "Confederacy of Ruined Lives"
-    assert tag.frame_set[b"XSOP"][0].text == "In the name of suffering"
-
-    # Convert to v2.4 and verify conversions
-    tag.save(version=ID3_V2_4)
-    audiofile = eyed3.load(audiofile.path)
-    tag = audiofile.tag
-    assert (set(tag.frame_set.keys()) ==
-            set([b"TPE1", b"TPE2", b"TSOP", b"TSO2"]))
-    assert tag.getTextFrame("TSO2") == "Dope sick"
-    assert tag.getTextFrame("TPE2") == "Take as needed for pain"
-    assert tag.getTextFrame("TPE1") == "Confederacy of Ruined Lives"
-    assert tag.getTextFrame("TSOP") == "In the name of suffering"
-
-
 def testSetNumExceptions():
     t = Tag()
     with pytest.raises(ValueError) as ex:
         t.track_num = (1, 2, 3)
-    with pytest.raises(TypeError) as ex:
-        t.track_num = "1"
+
+
+@deprecation.fail_if_not_removed
+def testNonStdGenre():
+    t = Tag()
+    t.non_std_genre = "Black Lips"
+    assert t.genre.id is None
+    assert t.genre.name == "Black Lips"
+
+
+def testNumStringConvert():
+    t = Tag()
+
+    t.track_num = "1"
+    assert t.track_num == (1, None)
+
+    t.disc_num = ("2", "6")
+    assert t.disc_num == (2, 6)
+
+
+def testReleaseDate_v23_v24():
+    """v23 does not have release date, only original release date."""
+    date = Date.parse("1980-07-03")
+    date2 = Date.parse("1926-07-05")
+    year = Date(1966)
+
+    tag = Tag()
+    assert tag.version == ID3_DEFAULT_VERSION
+
+    tag.version = ID3_V2_3
+    assert tag.version == ID3_V2_3
+
+    # Setting release date sets original release date
+    # v2.3 TORY get the year, XDOR get the full date; getter prefers XDOR
+    tag.release_date = "2020-03-08"
+    assert b"TORY" in tag.frame_set
+    assert b"XDOR" in tag.frame_set
+    assert tag.release_date == Date.parse("2020-03-08")
+    assert tag.original_release_date == Date(year=2020, month=3, day=8)
+
+    # Setting original release date sets release date
+    tag.original_release_date = year
+    assert tag.original_release_date == Date(1966)
+    assert tag.release_date == Date.parse("1966")
+    assert b"TORY" in tag.frame_set
+    # Year only value should clean up XDOR
+    assert b"XDOR" not in tag.frame_set
+
+    # Version convert to 2.4 converts original release date only
+    tag.release_date = date
+    assert b"TORY" in tag.frame_set
+    assert b"XDOR" in tag.frame_set
+    assert tag.original_release_date == date
+    assert tag.release_date == date
+    tag.version = ID3_V2_4
+    assert tag.original_release_date == date
+    assert tag.release_date is None
+
+    # v2.4 has both date types
+    tag.release_date = date2
+    assert tag.original_release_date == date
+    assert tag.release_date == date2
+    assert b"TORY" not in tag.frame_set
+    assert b"XDOR" not in tag.frame_set
+
+    # Convert back to 2.3 loses release date, only the year is copied to TORY
+    tag.version = ID3_V2_3
+    assert b"TORY" in tag.frame_set
+    assert b"XDOR" in tag.frame_set
+    assert tag.original_release_date == date
+    assert tag.release_date == Date.parse(str(date))